adicionando comentarios
