+++ Exercicio_4_segunda/src/exercicio/Main.java
@@ -3,8 +3,9 @@ package exercicio;
 public class Main {
 
 	public static void main(String[] args) {
-		Televisao television = new Televisao();
-		
-		television.menu();
+		//Creating class call
+		ControleRemoto cr = new ControleRemoto();
+		//Calling the menu
+		cr.menu();
 	}
 }
